requirements.txt
Flask
requests
python-whois
beautifulsoup4
lxml
